Split metadata module into a subdirectory
