+++ src/engine/strat_engine/backstore/metadata/mod.rs
@@ -1,0 +1,11 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+mod bda;
+mod mda;
+
+pub use self::{
+    bda::{StaticHeader, BDA},
+    mda::MDADataSize,
+};
